requirements.txt
annotated-types==0.7.0
anyio==4.6.0
async-timeout==4.0.3
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==43.0.0
exceptiongroup==1.2.2
fastapi==0.115.0
h11==0.14.0
idna==3.8
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
python-decouple==3.8
redis==5.1.0
requests==2.32.3
requests-pkcs12==1.25
sniffio==1.3.1
starlette==0.38.6
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.31.0
xmltodict==0.13.0
